Create LabelList.vue
Create a basic component that contains and displays the labels for a note.
Props:
 - canModify: Boolean
Whether the icons for removing, adding, and replacing labels appear

- school: Object
The object containing the data for the required school label

- course: Object
The object containing the data for the required course label

- labels: Array
All the remaining labels that have been submitted by users.