Add environment variables directly to config file.
This is not the most secure way to keep database credentials. I tried
separating the env_variables into a separate xml file and including it
in the main config but apparently that doesn't work for a maven project
on appengine. We can look into a more secure method of storing database
credentials in the future.