Add a context servlet that opens a connection pool to the database once the app starts.
In doing this, connection pools don't have to be reopened and closed
over and over again. By keeping the pool in the app context, the pool
will open once the app is started and simply close once the context is
destroyed i.e. the app closes. There is also a test servlet that can be
used to check if the DB connection is working.